main: Add support for debug local image in docker-compose
In some cases we want to test local builds on the emulator
infrastructure that this app provides. This commit adds support for
debugging local images in the docker-compose file. It adds a new
service called `emulator-debug` that runs the emulator in debug mode
and allows the user to pass a local image path as an argument.
This is done by checking if the `PHOBOS_LOCAL_IMG_PATH` environment
variable is set. If it is set, the `emulator-debug` service will be
used instead of the `emulator` service.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>